requirements.txt
flask
gunicorn
numpy==1.19.5
pandas==1.2.5
scikit-learn==1.0
matplotlib==3.3.4
PyYAML
-e .
